Simplifying the code by adding the overview of flashcards to own file
